+++ xarray/core/rolling.py
@@ -4,7 +4,7 @@ from __future__ import print_function
 import numpy as np
 
 from .pycompat import OrderedDict, zip
-from .common import ImplementsRollingArrayReduce, _full_like
+from .common import ImplementsRollingArrayReduce, full_like
 from .combine import concat
 from .ops import inject_bottleneck_rolling_methods
 
@@ -106,7 +106,7 @@ class Rolling(object):
             window = self.obj.isel(**{self.dim: indices})
 
             if not valid:
-                window = _full_like(window, fill_value=True)
+                window = full_like(window, fill_value=True, dtype=bool)
 
             yield (label, window)