+++ xarray/test/test_dataarray.py
@@ -4,6 +4,7 @@ from __future__ import print_function
 import numpy as np
 import pandas as pd
 import pickle
+import pytest
 from copy import deepcopy
 from textwrap import dedent
 
@@ -12,7 +13,7 @@ import xarray as xr
 from xarray import (align, broadcast, Dataset, DataArray,
                     IndexVariable, Variable)
 from xarray.core.pycompat import iteritems, OrderedDict
-from xarray.core.common import _full_like
+from xarray.core.common import full_like
 
 from xarray.test import (TestCase, ReturnItem, source_ndarray, unittest,
                          requires_dask, requires_bottleneck)
@@ -282,8 +283,9 @@ class TestDataArray(TestCase):
         with self.assertRaisesRegexp(ValueError, 'conflicting MultiIndex'):
             DataArray(np.random.rand(4, 4),
                       [('x', self.mindex), ('y', self.mindex)])
+        with self.assertRaisesRegexp(ValueError, 'conflicting MultiIndex'):
             DataArray(np.random.rand(4, 4),
-                      [('x', mindex), ('level_1', range(4))])
+                      [('x', self.mindex), ('level_1', range(4))])
 
     def test_constructor_from_self_described(self):
         data = [[-0.1, 21], [0, 2]]
@@ -869,6 +871,13 @@ class TestDataArray(TestCase):
             renamed.to_dataset(), self.ds.rename({'foo': 'bar'}))
         self.assertEqual(renamed.name, 'bar')
 
+        # regression to #1116
+        renamed = self.dv.x.rename({'x': 'z'})
+        self.assertDatasetIdentical(
+            renamed, self.ds.rename({'x': 'z'}).z)
+        self.assertEqual(renamed.name, 'z')
+        self.assertEqual(renamed.dims, ('z',))
+
     def test_swap_dims(self):
         array = DataArray(np.random.randn(3), {'y': ('x', list('abc'))}, 'x')
         expected = DataArray(array.values, {'y': list('abc')}, dims='y')
@@ -927,7 +936,6 @@ class TestDataArray(TestCase):
     def test_non_overlapping_dataarrays_return_empty_result(self):
 
         a = DataArray(range(5), [('x', range(5))])
-        b = DataArray(range(5), [('x', range(1, 6))])
         result = a.isel(x=slice(2)) + a.isel(x=slice(2, None))
         self.assertEqual(len(result['x']), 0)
 
@@ -936,7 +944,6 @@ class TestDataArray(TestCase):
         a = DataArray(data=[])
         result = a * a
         self.assertEqual(len(result['dim_0']), 0)
-
 
     def test_inplace_math_basics(self):
         x = self.x
@@ -2152,10 +2159,7 @@ class TestDataArray(TestCase):
         self.assertEqual(len(ma.mask), N)
 
     def test_to_and_from_cdms2(self):
-        try:
-            import cdms2
-        except ImportError:
-            raise unittest.SkipTest('cdms2 not installed')
+        pytest.importorskip('cdms2')
 
         original = DataArray(np.arange(6).reshape(2, 3),
                              [('distance', [-2, 2], {'units': 'meters'}),
@@ -2301,38 +2305,23 @@ class TestDataArray(TestCase):
             array.other = 2
 
     def test_full_like(self):
-        da = DataArray(np.random.random(size=(4, 4)), dims=('x', 'y'),
-                       attrs={'attr1': 'value1'})
-        actual = _full_like(da)
+        # For more thorough tests, see test_variable.py
+        da = DataArray(np.random.random(size=(2, 2)),
+                       dims=('x', 'y'),
+                       attrs={'attr1': 'value1'},
+                       coords={'x': [4, 3]},
+                       name='helloworld')
 
-        self.assertEqual(actual.dtype, da.dtype)
-        self.assertEqual(actual.shape, da.shape)
-        self.assertEqual(actual.dims, da.dims)
-        self.assertEqual(actual.attrs, {})
+        actual = full_like(da, 2)
+        expect = da.copy(deep=True)
+        expect.values = [[2.0, 2.0], [2.0, 2.0]]
+        self.assertDataArrayIdentical(expect, actual)
 
-        for name in da.coords:
-            self.assertArrayEqual(da[name], actual[name])
-            self.assertEqual(da[name].dtype, actual[name].dtype)
-
-        # keep attrs
-        actual = _full_like(da, keep_attrs=True)
-        self.assertEqual(actual.attrs, da.attrs)
-
-        # Fill value
-        actual = _full_like(da, fill_value=True)
-        self.assertEqual(actual.dtype, da.dtype)
-        self.assertEqual(actual.shape, da.shape)
-        self.assertEqual(actual.dims, da.dims)
-        np.testing.assert_equal(actual.values, np.nan)
-
-        actual = _full_like(da, fill_value=10)
-        self.assertEqual(actual.dtype, da.dtype)
-        np.testing.assert_equal(actual.values, 10)
-
-        # make sure filling with nans promotes integer type
-        actual = _full_like(DataArray([1, 2, 3]), fill_value=np.nan)
-        self.assertEqual(actual.dtype, np.float)
-        np.testing.assert_equal(actual.values, np.nan)
+        # override dtype
+        actual = full_like(da, fill_value=True, dtype=bool)
+        expect.values = [[True, True], [True, True]]
+        self.assertEquals(expect.dtype, bool)
+        self.assertDataArrayIdentical(expect, actual)
 
     def test_dot(self):
         x = np.linspace(-3, 3, 6)